created opt out apis and update data flow in service

+++ backend/scheduler-svc/src/main/java/digit/web/models/ReScheduleHearing.java
@@ -54,7 +54,7 @@ public class ReScheduleHearing {
     private AuditDetails auditDetails;
 
     @JsonProperty("rowVersion")
-    private Long rowVersion = null;
+    private Integer rowVersion = null;
 
     @JsonProperty("documents")
     @Valid
